Added Importer for Scenes

+++ Raytracing/Primitive.h
@@ -34,7 +34,7 @@ public:
 			Eigen::Vector3d(0.5,0.5,0.5),
 			Eigen::Vector3d(0.5,-0.5,0.5),
 		};
-		size_t triangles[36] =
+		int triangles[36] =
 		{
 			0,2,3,0,3,1,8,4,5,8,5,9,10,6,7,10,7,11,12,13,14,12,14,15,16,17,18,16,18,19,20,21,22,20,22,23
 		};
@@ -72,7 +72,7 @@ public:
 			Eigen::Vector3d(0.5,0.5,0.5),
 			Eigen::Vector3d(0.5,-0.5,0.5),
 		};
-		size_t triangles[36] =
+		int triangles[36] =
 		{
 			0,2,3,0,3,1,8,4,5,8,5,9,10,6,7,10,7,11,12,13,14,12,14,15,16,17,18,16,18,19,20,21,22,20,22,23
 		};
@@ -89,7 +89,7 @@ public:
 			Eigen::Vector3d(0.5, 0, -0.5),
 			Eigen::Vector3d(0.5, 0,  0.5)
 		};
-		size_t triangles[6] =
+		int triangles[6] =
 		{
 			0,1,2,3,2,1
 		};
@@ -106,7 +106,7 @@ public:
 			Eigen::Vector3d(0.5, 0, -0.5),
 			Eigen::Vector3d(0.5, 0,  0.5)
 		};
-		size_t triangles[6] =
+		int triangles[6] =
 		{
 			0,1,2,3,2,1
 		};
@@ -125,7 +125,7 @@ public:
 			Eigen::Vector3d(-0.5, 0  ,  0.5),	// v3
 			Eigen::Vector3d(0  , 1,    0)		// v4
 		};
-		size_t triangles[18] =
+		int triangles[18] =
 		{
 			0,1,4,
 			1,2,4,
@@ -148,7 +148,7 @@ public:
 			Eigen::Vector3d(-0.5, 0  ,  0.5),	// v3
 			Eigen::Vector3d(0  , 1,    0)		// v4
 		};
-		size_t triangles[18] =
+		int triangles[18] =
 		{
 			0,1,4,
 			1,2,4,
